Fix parts of CGP extension that could not longer build against recent versions of the ECJ public interface.

+++ contrib/cgp/ec/cgp/ESSelectionCGP.java
@@ -48,7 +48,7 @@ public class ESSelectionCGP extends ESSelection {
 		 * random.
 		 */
 		if (pos == 0) { /* only do this for the first parent. */
-			float fit = inds2[parent].fitness.fitness();
+			double fit = inds2[parent].fitness.fitness();
 			int x = 0;
 			while (fit == inds2[parent + x].fitness.fitness()
 					&& parent + x < inds.length)
